Multiple kegtron sensors
