+++ custom_components/ble_monitor/__init__.py
@@ -3,11 +3,8 @@ import asyncio
 import copy
 import json
 import logging
-import math
-import struct
 from threading import Thread
 import janus
-from Cryptodome.Cipher import AES
 import voluptuous as vol
 
 from homeassistant.config_entries import SOURCE_IMPORT, ConfigEntry
@@ -31,6 +28,7 @@ from homeassistant.helpers.entity_registry import (
 # the official aioblescan repo see https://github.com/frawau/aioblescan/pull/30,
 # thanks to @vicamo
 from . import aioblescan_ext as aiobs
+from .ble_parser import ble_parser
 
 from .const import (
     DEFAULT_ROUNDING,
@@ -68,32 +66,10 @@ from .const import (
     PLATFORMS,
     MAC_REGEX,
     AES128KEY_REGEX,
-    ATC_TYPE_DICT,
-    QINGPING_TYPE_DICT,
-    XIAOMI_TYPE_DICT,
-    MISCALE_TYPE_DICT,
-    KEGTRON_TYPE_DICT,
     SERVICE_CLEANUP_ENTRIES,
 )
 
 _LOGGER = logging.getLogger(__name__)
-
-# Structured objects for data conversions
-TH_STRUCT = struct.Struct("<hH")
-H_STRUCT = struct.Struct("<H")
-T_STRUCT = struct.Struct("<h")
-TTB_STRUCT = struct.Struct("<hhB")
-CND_STRUCT = struct.Struct("<H")
-ILL_STRUCT = struct.Struct("<I")
-LIGHT_STRUCT = struct.Struct("<I")
-FMDH_STRUCT = struct.Struct("<H")
-THBV_STRUCT = struct.Struct(">hBBH")
-THVB_STRUCT = struct.Struct("<hHHB")
-M_STRUCT = struct.Struct("<L")
-P_STRUCT = struct.Struct("<H")
-SCALE_V1_STRUCT = struct.Struct("<BH7x")
-SCALE_V2_STRUCT = struct.Struct("<BB7xHH")
-KEGTRON_STRUCT = struct.Struct(">HHHB20x")
 
 CONFIG_YAML = {}
 UPDATE_UNLISTENER = None
@@ -454,272 +430,6 @@ class HCIdump(Thread):
     def __init__(self, config, dataqueue):
         """Initiate HCIdump thread."""
 
-        # Xiaomi MiBeacon BLE advertisements
-        # https://iot.mi.com/new/doc/embedded-development/ble/object-definition
-        def obj0300(xobj):
-            return {"motion": xobj[0], "motion timer": xobj[0]}
-
-        def obj0f00(xobj):
-            if len(xobj) == 3:
-                (value,) = LIGHT_STRUCT.unpack(xobj + b'\x00')
-                # MJYD02YL:  1 - moving no light, 100 - moving with light
-                # RTCGQ02LM: 0 - moving no light, 256 - moving with light
-                # CGPR1:     moving, value is illumination in lux
-                return {"motion": 1, "motion timer": 1, "light": int(value >= 100), "illuminance": value}
-            else:
-                return {}
-
-        def obj0110(xobj):
-            if xobj[2] == 0:
-                press = "single press"
-            elif xobj[2] == 1:
-                press = "double press"
-            elif xobj[2] == 2:
-                press = "long press"
-            else:
-                press = "no press"
-            return {"button": press}
-
-        def obj0410(xobj):
-            if len(xobj) == 2:
-                (temp,) = T_STRUCT.unpack(xobj)
-                return {"temperature": temp / 10}
-            else:
-                return {}
-
-        def obj0510(xobj):
-            return {"switch": xobj[0], "temperature": xobj[1]}
-
-        def obj0610(xobj):
-            if len(xobj) == 2:
-                (humi,) = H_STRUCT.unpack(xobj)
-                return {"humidity": humi / 10}
-            else:
-                return {}
-
-        def obj0710(xobj):
-            if len(xobj) == 3:
-                (illum,) = ILL_STRUCT.unpack(xobj + b'\x00')
-                return {"illuminance": illum, "light": 1 if illum == 100 else 0}
-            else:
-                return {}
-
-        def obj0810(xobj):
-            return {"moisture": xobj[0]}
-
-        def obj0910(xobj):
-            if len(xobj) == 2:
-                (cond,) = CND_STRUCT.unpack(xobj)
-                return {"conductivity": cond}
-            else:
-                return {}
-
-        def obj1010(xobj):
-            if len(xobj) == 2:
-                (fmdh,) = FMDH_STRUCT.unpack(xobj)
-                return {"formaldehyde": fmdh / 100}
-            else:
-                return {}
-
-        def obj1210(xobj):
-            return {"switch": xobj[0]}
-
-        def obj1310(xobj):
-            return {"consumable": xobj[0]}
-
-        def obj1410(xobj):
-            return {"moisture": xobj[0]}
-
-        def obj1710(xobj):
-            if len(xobj) == 4:
-                (motion,) = M_STRUCT.unpack(xobj)
-                # seconds since last motion detected message (not used, we use motion timer in obj0f00)
-                # 0 = motion detected
-                return {"motion": 1 if motion == 0 else 0}
-            else:
-                return {}
-
-        def obj1810(xobj):
-            return {"light": xobj[0]}
-
-        def obj1910(xobj):
-            return {"opening": xobj[0]}
-
-        def obj0a10(xobj):
-            return {"battery": xobj[0]}
-
-        def obj0d10(xobj):
-            if len(xobj) == 4:
-                (temp, humi) = TH_STRUCT.unpack(xobj)
-                return {"temperature": temp / 10, "humidity": humi / 10}
-            else:
-                return {}
-
-        def obj0020(xobj):
-            if len(xobj) == 5:
-                (temp1, temp2, bat) = TTB_STRUCT.unpack(xobj)
-                # Body temperature is calculated from the two measured temperatures.
-                # Formula is based on approximation based on values inthe app in the range 36.5 - 37.8.
-                body_temp = (
-                    3.71934 * pow(10, -11) * math.exp(0.69314 * temp1 / 100)
-                    - 1.02801 * pow(10, -8) * math.exp(0.53871 * temp2 / 100)
-                    + 36.413
-                )
-                return {"temperature": body_temp, "battery": bat}
-            else:
-                return {}
-
-        # Qingping BLE advertisements
-        def obj0104(xobj):
-            if len(xobj) == 4:
-                (temp, humi) = TH_STRUCT.unpack(xobj)
-                return {"temperature": temp / 10, "humidity": humi / 10}
-            else:
-                return {}
-
-        def obj0201(xobj):
-            return {"battery": xobj[0]}
-
-        def obj0702(xobj):
-            if len(xobj) == 2:
-                (pres,) = P_STRUCT.unpack(xobj)
-                return {"pressure": pres / 10}
-            else:
-                return {}
-
-        # ATC BLE advertisements
-        def objATC_short(xobj):
-            if len(xobj) == 6:
-                (temp, humi, batt, volt) = THBV_STRUCT.unpack(xobj)
-                return {"temperature": temp / 10, "humidity": humi, "voltage": volt / 1000, "battery": batt}
-            else:
-                return {}
-
-        def objATC_long(xobj):
-            if len(xobj) == 7:
-                (temp, humi, volt, batt) = THVB_STRUCT.unpack(xobj)
-                return {"temperature": temp / 100, "humidity": humi / 100, "voltage": volt / 1000, "battery": batt}
-            else:
-                return {}
-
-        # MI Scale V1 BLE advertisements
-        def obj1d18(xobj):
-            if len(xobj) == 10:
-                (controlByte, weight) = SCALE_V1_STRUCT.unpack(xobj)
-                isStabilized = controlByte & (1 << 5)
-                weightRemoved = controlByte & (1 << 7)
-
-                if controlByte & (1 << 0):
-                    weight = weight / 100
-                    weight_unit = 'lbs'
-                elif controlByte & (1 << 4):
-                    weight = weight / 100
-                    weight_unit = 'jin'
-                else:
-                    weight = weight / 200
-                    weight_unit = 'kg'
-
-                if isStabilized:
-                    return {
-                        "weight": weight,
-                        "non-stabilized weight": weight,
-                        "weight unit": weight_unit,
-                        "weight removed": 0 if weightRemoved == 0 else 1,
-                        "stabilized": 0 if isStabilized == 0 else 1
-                    }
-                else:
-                    return {
-                        "non-stabilized weight": weight,
-                        "weight unit": weight_unit,
-                        "weight removed": 0 if weightRemoved == 0 else 1,
-                        "stabilized": 0 if isStabilized == 0 else 1
-                    }
-            else:
-                return {}
-
-        # MI Scale V2 BLE advertisements
-        def obj1b18(xobj):
-            if len(xobj) == 13:
-                (measunit, controlByte, impedance, weight) = SCALE_V2_STRUCT.unpack(xobj)
-                hasImpedance = controlByte & (1 << 1)
-                isStabilized = controlByte & (1 << 5)
-                weightRemoved = controlByte & (1 << 7)
-
-                if measunit & (1 << 4):
-                    # measurement in Chinese Catty unit
-                    weight = weight / 100
-                    weight_unit = "jin"
-                elif measunit == 3:
-                    # measurement in lbs
-                    weight = weight / 100
-                    weight_unit = "lbs"
-                elif measunit == 2:
-                    # measurement in kg
-                    weight = weight / 200
-                    weight_unit = "kg"
-                else:
-                    # measurement in unknown unit
-                    weight = weight / 100
-                    weight_unit = None
-
-                if isStabilized:
-                    if hasImpedance:
-                        return {
-                            "weight": weight,
-                            "non-stabilized weight": weight,
-                            "weight unit": weight_unit,
-                            "impedance": impedance,
-                            "weight removed": 0 if weightRemoved == 0 else 1,
-                            "stabilized": 0 if isStabilized == 0 else 1
-                        }
-                    else:
-                        return {
-                            "weight": weight,
-                            "non-stabilized weight": weight,
-                            "weight unit": weight_unit,
-                            "weight removed": 0 if weightRemoved == 0 else 1,
-                            "stabilized": 0 if isStabilized == 0 else 1
-                        }
-                else:
-                    return {
-                        "non-stabilized weight": weight,
-                        "weight unit": weight_unit,
-                        "weight removed": 0 if weightRemoved == 0 else 1,
-                        "stabilized": 0 if isStabilized == 0 else 1
-                    }
-            else:
-                return {}
-
-        # Kegtron BLE advertisements
-        def objKegtron(xobj):
-            if len(xobj) == 27:
-                (keg_size, vol_start, vol_disp, port) = KEGTRON_STRUCT.unpack(xobj)
-                if port & (1 << 0):
-                    port_state = "configured"
-                else:
-                    port_state = "unconfigured (new device)"
-
-                if port & (1 << 4):
-                    port_index = "port 2"
-                else:
-                    port_index = "port 1"
-
-                if port & (1 << 6):
-                    port_count = "Dual port device"
-                else:
-                    port_count = "Single port device"
-
-                return {
-                    "keg size": keg_size,
-                    "volume start": vol_start,
-                    "volume dispensed": vol_disp,
-                    "port state": port_state,
-                    "port index": port_index,
-                    "port count": port_count
-                }
-            else:
-                return {}
-
         def reverse_mac(rmac):
             """Change LE order to BE."""
             if len(rmac) != 12:
@@ -771,44 +481,13 @@ class HCIdump(Thread):
         for i, mac in enumerate(self.whitelist):
             self.whitelist[i] = bytes.fromhex(reverse_mac(mac.replace(":", "")).lower())
         _LOGGER.debug("%s whitelist item(s) loaded.", len(self.whitelist))
-        # dataobject dictionary to implement switch-case statement
-        # dataObject id  (converter, binary, measuring)
-        self._dataobject_dict = {
-            b'\x03\x00': (obj0300, True, False),
-            b'\x0F\x00': (obj0f00, True, True),
-            b'\x01\x10': (obj0110, False, True),
-            b'\x04\x10': (obj0410, False, True),
-            b'\x05\x10': (obj0510, True, True),
-            b'\x06\x10': (obj0610, False, True),
-            b'\x07\x10': (obj0710, True, True),
-            b'\x08\x10': (obj0810, False, True),
-            b'\x09\x10': (obj0910, False, True),
-            b'\x10\x10': (obj1010, False, True),
-            b'\x12\x10': (obj1210, True, False),
-            b'\x13\x10': (obj1310, False, True),
-            b'\x14\x10': (obj1410, True, False),
-            b'\x17\x10': (obj1710, True, False),
-            b'\x18\x10': (obj1810, True, False),
-            b'\x19\x10': (obj1910, True, False),
-            b'\x0A\x10': (obj0a10, True, True),
-            b'\x0D\x10': (obj0d10, False, True),
-            b'\x00\x20': (obj0020, False, True),
-            b'\x01\x04': (obj0104, False, True),
-            b'\x02\x01': (obj0201, False, True),
-            b'\x07\x02': (obj0702, False, True),
-            b'\x10\x16': (objATC_short, False, True),
-            b'\x12\x16': (objATC_long, False, True),
-            b'\x1B\x18': (obj1b18, True, True),
-            b'\x1D\x18': (obj1d18, True, True),
-            b'\xFF\xFF': (objKegtron, False, True),
-        }
 
     def process_hci_events(self, data):
         """Parse HCI events."""
         self.evt_cnt += 1
         if len(data) < 12:
             return
-        msg, binary, measuring = self.parse_raw_message(data)
+        msg, binary, measuring = ble_parser(self, data)
         if msg:
             if binary == measuring:
                 self.dataqueue_bin.sync_q.put_nowait(msg)
@@ -891,683 +570,3 @@ class HCIdump(Thread):
             self._event_loop.call_soon_threadsafe(self._event_loop.stop)
         except AttributeError as error:
             _LOGGER.debug("%s", error)
-
-    def parse_raw_message(self, data):
-        """Parse the raw data."""
-
-        # check if packet is Extended scan result
-        is_ext_packet = True if data[3] == 0x0d else False
-
-        # check for service data (Xiaomi, qingping or ATC)
-        xiaomi_index = data.find(b'\x16\x95\xFE', 15 + 15 if is_ext_packet else 0)
-        qingping_index = data.find(b'\x16\xCD\xFD', 15 + 15 if is_ext_packet else 0)
-        atc_index = data.find(b'\x16\x1A\x18', 15 + 15 if is_ext_packet else 0)
-        miscale_v1_index = data.find(b'\x16\x1D\x18', 15 + 15 if is_ext_packet else 0)
-        miscale_v2_index = data.find(b'\x16\x1B\x18', 15 + 15 if is_ext_packet else 0)
-        kegtron_index = data.find(b'\x1E\xFF\xFF\xFF', 14 + 15 if is_ext_packet else 0)
-
-        try:
-            if xiaomi_index != -1:
-                return self.parse_xiaomi(data, xiaomi_index, is_ext_packet)
-            elif qingping_index != -1:
-                return self.parse_qingping(data, qingping_index, is_ext_packet)
-            elif atc_index != -1:
-                return self.parse_atc(data, atc_index, is_ext_packet)
-            elif miscale_v1_index != -1:
-                return self.parse_miscale(data, miscale_v1_index, is_ext_packet)
-            elif miscale_v2_index != -1:
-                return self.parse_miscale(data, miscale_v2_index, is_ext_packet)
-            elif kegtron_index != -1:
-                return self.parse_kegtron(data, kegtron_index, is_ext_packet)
-
-        except NoValidError as nve:
-            _LOGGER.debug("Invalid data: %s", nve)
-
-        return None, None, None
-
-    def parse_xiaomi(self, data, xiaomi_index, is_ext_packet):
-        # parse BLE message in Xiaomi MiBeacon format
-        firmware = "Xiaomi (MiBeacon)"
-
-        # check for no BR/EDR + LE General discoverable mode flags
-        advert_start = 29 if is_ext_packet else 14
-        adv_index = data.find(b"\x02\x01\x06", advert_start, 3 + advert_start)
-        adv_index2 = data.find(b"\x15\x16\x95", advert_start, 3 + advert_start)
-        if adv_index == -1 and adv_index2 == -1:
-            raise NoValidError("Invalid index")
-        if adv_index2 != -1:
-            adv_index = adv_index2
-
-        # check for BTLE msg size
-        msg_length = data[2] + 3
-        if msg_length != len(data):
-            raise NoValidError("Invalid msg size")
-
-        # extract device type
-        device_type = data[xiaomi_index + 5:xiaomi_index + 7]
-
-        # extract frame control bits
-        framectrl_data = data[xiaomi_index + 3:xiaomi_index + 5]
-        framectrl, = struct.unpack('>H', framectrl_data)
-
-        # flag advertisements without mac address in service data
-        if device_type == b'\xF6\x07' and framectrl_data == b'\x48\x59':
-            # MJYD02YL does not have a MAC address in the service data of some advertisements
-            mac_in_service_data = False
-        elif device_type == b'\xDD\x03' and framectrl_data == b'\x40\x30':
-            # MUE4094RT does not have a MAC address in the service data
-            mac_in_service_data = False
-        else:
-            mac_in_service_data = True
-
-        # check for MAC presence in message and in service data
-        mac_index = adv_index - 14 if is_ext_packet else adv_index
-        if mac_in_service_data is True:
-            xiaomi_mac_reversed = data[xiaomi_index + 8:xiaomi_index + 14]
-            source_mac_reversed = data[mac_index - 7:mac_index - 1]
-            if xiaomi_mac_reversed != source_mac_reversed:
-                raise NoValidError("Invalid MAC address")
-        else:
-            # for sensors without mac in service data, use the first mac in advertisment
-            xiaomi_mac_reversed = data[mac_index - 7:mac_index - 1]
-
-        # check for MAC presence in whitelist, if needed
-        if self.discovery is False and xiaomi_mac_reversed not in self.whitelist:
-            return None, None, None
-        packet_id = data[xiaomi_index + 7]
-        try:
-            prev_packet = self.lpacket_ids[xiaomi_mac_reversed]
-        except KeyError:
-            # start with empty first packet
-            prev_packet = None, None, None
-        if prev_packet == packet_id:
-            # only process new messages
-            return None, None, None
-        self.lpacket_ids[xiaomi_mac_reversed] = packet_id
-
-        # extract RSSI byte
-        rssi_index = 18 if is_ext_packet else msg_length - 1
-        (rssi,) = struct.unpack("<b", data[rssi_index:rssi_index + 1])
-
-        # strange positive RSSI workaround
-        if rssi > 0:
-            rssi = -rssi
-        try:
-            sensor_type, binary_data = XIAOMI_TYPE_DICT[device_type]
-        except KeyError:
-            if self.report_unknown:
-                _LOGGER.info(
-                    "BLE ADV from UNKNOWN: RSSI: %s, MAC: %s, ADV: %s",
-                    rssi,
-                    ''.join('{:02X}'.format(x) for x in xiaomi_mac_reversed[::-1]),
-                    data.hex()
-                )
-            raise NoValidError("Device unkown")
-
-        # check data is present
-        if not (framectrl & 0x4000):
-            return {
-                "rssi": rssi,
-                "mac": ''.join('{:02X}'.format(x) for x in xiaomi_mac_reversed[::-1]),
-                "type": sensor_type,
-                "packet": packet_id,
-                "firmware": firmware,
-                "data": False,
-            }, None, None
-        xdata_length = 0
-        xdata_point = 0
-
-        # check capability byte present
-        if framectrl & 0x2000:
-            xdata_length = -1
-            xdata_point = 1
-
-        # check for messages without mac address in service data
-        if mac_in_service_data is False:
-            xdata_length = +6
-            xdata_point = -6
-
-        # parse_xiaomi data length = message length
-        #     -all bytes before XiaomiUUID
-        #     -3 bytes Xiaomi UUID + ADtype
-        #     -1 byte rssi
-        #     -3+1 bytes sensor type
-        #     -1 byte packet_id
-        #     -6 bytes MAC (if present)
-        #     -capability byte offset
-        xdata_length += msg_length - xiaomi_index - 15
-        if xdata_length < 3:
-            raise NoValidError("Xdata length invalid")
-
-        xdata_point += xiaomi_index + 14
-
-        # check if parse_xiaomi data start and length is valid
-        if xdata_length != len(data[xdata_point:-1]):
-            raise NoValidError("Invalid data length")
-
-        # check encrypted data flags
-        if framectrl & 0x0800:
-            # try to find encryption key for current device
-            try:
-                key = self.aeskeys[xiaomi_mac_reversed]
-            except KeyError:
-                # no encryption key found
-                raise NoValidError("No encryption key found")
-            nonce = b"".join(
-                [
-                    xiaomi_mac_reversed,
-                    device_type,
-                    data[xiaomi_index + 7:xiaomi_index + 8]
-                ]
-            )
-            endoffset = msg_length - int(not is_ext_packet)
-            encrypted_payload = data[xdata_point:endoffset]
-            aad = b"\x11"
-            token = encrypted_payload[-4:]
-            payload_counter = encrypted_payload[-7:-4]
-            nonce = b"".join([nonce, payload_counter])
-            cipherpayload = encrypted_payload[:-7]
-            cipher = AES.new(key, AES.MODE_CCM, nonce=nonce, mac_len=4)
-            cipher.update(aad)
-
-            try:
-                decrypted_payload = cipher.decrypt_and_verify(cipherpayload, token)
-            except ValueError as error:
-                _LOGGER.error("Decryption failed: %s", error)
-                _LOGGER.error("token: %s", token.hex())
-                _LOGGER.error("nonce: %s", nonce.hex())
-                _LOGGER.error("encrypted_payload: %s", encrypted_payload.hex())
-                _LOGGER.error("cipherpayload: %s", cipherpayload.hex())
-                raise NoValidError("Error decrypting with arguments")
-            if decrypted_payload is None:
-                _LOGGER.error(
-                    "Decryption failed for %s, decrypted payload is None",
-                    "".join("{:02X}".format(x) for x in xiaomi_mac_reversed[::-1]),
-                )
-                raise NoValidError("Decryption failed")
-
-            # replace cipher with decrypted data
-            msg_length -= len(encrypted_payload)
-            if is_ext_packet:
-                data = b"".join((data[:xdata_point], decrypted_payload))
-            else:
-                data = b"".join((data[:xdata_point], decrypted_payload, data[-1:]))
-            msg_length += len(decrypted_payload)
-
-        result = {
-            "rssi": rssi,
-            "mac": ''.join('{:02X}'.format(x) for x in xiaomi_mac_reversed[::-1]),
-            "type": sensor_type,
-            "packet": packet_id,
-            "firmware": firmware,
-            "data": True,
-        }
-        binary = False
-        measuring = False
-
-        # loop through parse_xiaomi payload
-        # assume that the data may have several values of different types,
-        # although I did not notice this behavior with my LYWSDCGQ sensors
-        while True:
-            xvalue_typecode = data[xdata_point:xdata_point + 2]
-            try:
-                xvalue_length = data[xdata_point + 2]
-            except ValueError as error:
-                _LOGGER.error("xvalue_length conv. error: %s", error)
-                _LOGGER.error("xdata_point: %s", xdata_point)
-                _LOGGER.error("data: %s", data.hex())
-                result = {}
-                break
-            except IndexError as error:
-                _LOGGER.error("Wrong xdata_point: %s", error)
-                _LOGGER.error("xdata_point: %s", xdata_point)
-                _LOGGER.error("data: %s", data.hex())
-                result = {}
-                break
-
-            xnext_point = xdata_point + 3 + xvalue_length
-            xvalue = data[xdata_point + 3:xnext_point]
-            resfunc, tbinary, tmeasuring = self._dataobject_dict.get(xvalue_typecode, (None, None, None))
-
-            if resfunc:
-                binary = binary or tbinary
-                measuring = measuring or tmeasuring
-                result.update(resfunc(xvalue))
-            else:
-                if self.report_unknown:
-                    _LOGGER.info(
-                        "UNKNOWN dataobject from DEVICE: %s, MAC: %s, ADV: %s",
-                        sensor_type,
-                        ''.join('{:02X}'.format(x) for x in xiaomi_mac_reversed[::-1]),
-                        data.hex()
-                    )
-
-            if xnext_point > msg_length - 3:
-                break
-            xdata_point = xnext_point
-
-        binary = binary and binary_data
-        return result, binary, measuring
-
-    def parse_qingping(self, data, qingping_index, is_ext_packet):
-        # parse BLE message in Qingping format
-        firmware = "Qingping"
-
-        # check for no BR/EDR + LE General discoverable mode flags
-        advert_start = 29 if is_ext_packet else 14
-        adv_index = data.find(b"\x02\x01\x06", advert_start, 3 + advert_start)
-        if adv_index == -1:
-            raise NoValidError("Invalid index")
-
-        # check for BTLE msg size
-        msg_length = data[2] + 3
-        if msg_length != len(data):
-            raise NoValidError("Invalid msg size")
-
-        # extract device type
-        device_type = data[qingping_index + 3:qingping_index + 5]
-
-        # check for MAC presence in message and in service data
-        mac_index = adv_index - 14 if is_ext_packet else adv_index
-        qingping_mac_reversed = data[qingping_index + 5:qingping_index + 11]
-        source_mac_reversed = data[mac_index - 7:mac_index - 1]
-        if qingping_mac_reversed != source_mac_reversed:
-            raise NoValidError("Invalid MAC address")
-
-        # check for MAC presence in whitelist, if needed
-        if self.discovery is False and qingping_mac_reversed not in self.whitelist:
-            return None, None, None
-        packet_id = "no packed id"
-
-        # extract RSSI byte
-        rssi_index = 18 if is_ext_packet else msg_length - 1
-        (rssi,) = struct.unpack("<b", data[rssi_index:rssi_index + 1])
-
-        # strange positive RSSI workaround
-        if rssi > 0:
-            rssi = -rssi
-        try:
-            sensor_type, binary_data = QINGPING_TYPE_DICT[device_type]
-        except KeyError:
-            if self.report_unknown:
-                _LOGGER.info(
-                    "BLE ADV from UNKNOWN: RSSI: %s, MAC: %s, ADV: %s",
-                    rssi,
-                    ''.join('{:02X}'.format(x) for x in qingping_mac_reversed[::-1]),
-                    data.hex()
-                )
-            raise NoValidError("Device unkown")
-        xdata_length = 0
-        xdata_point = 0
-
-        # parse_qingping data length = message length
-        #     -all bytes before Qingping UUID
-        #     -3 bytes Qingping UUID + ADtype
-        #     -1 byte rssi
-        #     -2 bytes sensor type
-        #     -6 bytes MAC
-        xdata_length += msg_length - qingping_index - 12
-        if xdata_length < 3:
-            raise NoValidError("Xdata length invalid")
-
-        xdata_point += qingping_index + 11
-
-        # check if parse_qingping data start and length is valid
-        if xdata_length != len(data[xdata_point:-1]):
-            raise NoValidError("Invalid data length")
-        result = {
-            "rssi": rssi,
-            "mac": ''.join('{:02X}'.format(x) for x in qingping_mac_reversed[::-1]),
-            "type": sensor_type,
-            "packet": packet_id,
-            "firmware": firmware,
-            "data": True,
-        }
-        binary = False
-        measuring = False
-
-        # loop through parse_qingping payload
-        # assume that the data may have several values of different types
-        while True:
-            xvalue_typecode = data[xdata_point:xdata_point + 2]
-            try:
-                xvalue_length = data[xdata_point + 1]
-            except ValueError as error:
-                _LOGGER.error("xvalue_length conv. error: %s", error)
-                _LOGGER.error("xdata_point: %s", xdata_point)
-                _LOGGER.error("data: %s", data.hex())
-                result = {}
-                break
-            except IndexError as error:
-                _LOGGER.error("Wrong xdata_point: %s", error)
-                _LOGGER.error("xdata_point: %s", xdata_point)
-                _LOGGER.error("data: %s", data.hex())
-                result = {}
-                break
-            xnext_point = xdata_point + 2 + xvalue_length
-            xvalue = data[xdata_point + 2:xnext_point]
-            resfunc, tbinary, tmeasuring = self._dataobject_dict.get(xvalue_typecode, (None, None, None))
-            if resfunc:
-                binary = binary or tbinary
-                measuring = measuring or tmeasuring
-                result.update(resfunc(xvalue))
-            else:
-                if self.report_unknown:
-                    _LOGGER.info(
-                        "UNKNOWN dataobject from DEVICE: %s, MAC: %s, ADV: %s",
-                        sensor_type,
-                        ''.join('{:02X}'.format(x) for x in qingping_mac_reversed[::-1]),
-                        data.hex()
-                    )
-            if xnext_point > msg_length - 3:
-                break
-            xdata_point = xnext_point
-        binary = binary and binary_data
-        return result, binary, measuring
-
-    def parse_atc(self, data, atc_index, is_ext_packet):
-
-        # parse BLE message in ATC format
-        # Check for the atc1441 or custom format
-        is_custom_adv = True if data[atc_index - 1] == 18 else False
-        if is_custom_adv:
-            firmware = "ATC firmware (custom)"
-        else:
-            firmware = "ATC firmware (ATC1441)"
-        # Check for old format (ATC firmware <= 2.8)
-        old_format = True if data.find(b"\x02\x01\x06", atc_index - 4, atc_index - 1) == -1 else False
-
-        # check for BTLE msg size
-        msg_length = data[2] + 3
-        if msg_length != len(data):
-            raise NoValidError("Invalid index")
-
-        # check for MAC presence in message and in service data
-        if is_custom_adv is True:
-            atc_mac_reversed = data[atc_index + 3:atc_index + 9]
-            atc_mac = atc_mac_reversed[::-1]
-        else:
-            atc_mac = data[atc_index + 3:atc_index + 9]
-
-        mac_index = atc_index - (22 if is_ext_packet else 8) - (0 if old_format else 3)
-        source_mac_reversed = data[mac_index:mac_index + 6]
-        source_mac = source_mac_reversed[::-1]
-        if atc_mac != source_mac:
-            raise NoValidError("Invalid MAC address")
-
-        # check for MAC presence in whitelist, if needed
-        if self.discovery is False and source_mac_reversed not in self.whitelist:
-            return None, None, None
-
-        packet_id = data[atc_index + 16 if is_custom_adv else atc_index + 15]
-        try:
-            prev_packet = self.lpacket_ids[atc_index]
-        except KeyError:
-            # start with empty first packet
-            prev_packet = None, None, None
-        if prev_packet == packet_id:
-            # only process new messages
-            return None, None, None
-        self.lpacket_ids[atc_index] = packet_id
-
-        # extract RSSI byte
-        rssi_index = 18 if is_ext_packet else msg_length - 1
-        (rssi,) = struct.unpack("<b", data[rssi_index:rssi_index + 1])
-
-        # strange positive RSSI workaround
-        if rssi > 0:
-            rssi = -rssi
-        device_type = data[atc_index + 1:atc_index + 3]
-        try:
-            sensor_type, binary_data = ATC_TYPE_DICT[device_type]
-        except KeyError:
-            if self.report_unknown:
-                _LOGGER.info(
-                    "BLE ADV from UNKNOWN ATC SENSOR: RSSI: %s, MAC: %s, ADV: %s",
-                    rssi,
-                    ''.join('{:02X}'.format(x) for x in atc_mac[:]),
-                    data.hex()
-                )
-            raise NoValidError("Device unkown")
-
-        # ATC data length = message length
-        # -all bytes before ATC UUID
-        # -3 bytes ATC UUID + ADtype
-        # -6 bytes MAC
-        # -1 Frame packet counter
-        # -1 byte flags (custom adv only)
-        # -1 RSSI (normal, not extended packet only)
-        xdata_length = msg_length - atc_index - (11 if is_custom_adv else 10) - (0 if is_ext_packet else 1)
-        if xdata_length < 6:
-            raise NoValidError("Xdata length invalid")
-
-        xdata_point = atc_index + 9
-
-        # check if parse_atc data start and length is valid
-        xdata_end_offset = (-1 if is_ext_packet else -2) + (-1 if is_custom_adv else 0)
-        if xdata_length != len(data[xdata_point:xdata_end_offset]):
-            raise NoValidError("Invalid data length")
-
-        result = {
-            "rssi": rssi,
-            "mac": ''.join('{:02X}'.format(x) for x in atc_mac[:]),
-            "type": sensor_type,
-            "packet": packet_id,
-            "firmware": firmware,
-            "data": True,
-        }
-        binary = False
-        measuring = False
-        xvalue_typecode = data[atc_index - 1:atc_index + 1]
-        xnext_point = xdata_point + xdata_length
-        xvalue = data[xdata_point:xnext_point]
-        resfunc, tbinary, tmeasuring = self._dataobject_dict.get(xvalue_typecode, (None, None, None))
-        if resfunc:
-            binary = binary or tbinary
-            measuring = measuring or tmeasuring
-            result.update(resfunc(xvalue))
-        else:
-            if self.report_unknown:
-                _LOGGER.info(
-                    "UNKNOWN dataobject from ATC DEVICE: %s, MAC: %s, ADV: %s",
-                    sensor_type,
-                    ''.join('{:02X}'.format(x) for x in atc_mac[:]),
-                    data.hex()
-                )
-        binary = binary and binary_data
-        return result, binary, measuring
-
-    def parse_miscale(self, data, miscale_index, is_ext_packet):
-        # parse BLE message in MiScale (v1 or v2) format
-
-        # check for no BR/EDR + LE General discoverable mode flags
-        advert_start = 29 if is_ext_packet else 14
-        adv_index = data.find(b"\x02\x01\x06", advert_start, 3 + advert_start)
-        if adv_index == -1:
-            raise NoValidError("Invalid index")
-
-        # check for BTLE msg size
-        msg_length = data[2] + 3
-        if msg_length != len(data):
-            raise NoValidError("Invalid msg size")
-
-        # extract device type
-        device_type = data[miscale_index + 5:miscale_index + 7]
-
-        # check for MAC presence in message and in service data
-        mac_index = adv_index - 14 if is_ext_packet else adv_index
-        miscale_mac_reversed = data[mac_index - 7:mac_index - 1]
-        miscale_mac = miscale_mac_reversed[::-1]
-
-        # check for MAC presence in whitelist, if needed
-        if self.discovery is False and miscale_mac not in self.whitelist:
-            return None, None, None
-
-        # extract RSSI byte
-        rssi_index = 18 if is_ext_packet else msg_length - 1
-        (rssi,) = struct.unpack("<b", data[rssi_index:rssi_index + 1])
-
-        # strange positive RSSI workaround
-        if rssi > 0:
-            rssi = -rssi
-        device_type = data[miscale_index + 1:miscale_index + 3]
-        try:
-            sensor_type, binary_data = MISCALE_TYPE_DICT[device_type]
-        except KeyError:
-            if self.report_unknown:
-                _LOGGER.info(
-                    "BLE ADV from UNKNOWN MI SCALE SENSOR: RSSI: %s, MAC: %s, ADV: %s",
-                    rssi,
-                    ''.join('{:02X}'.format(x) for x in miscale_mac[:]),
-                    data.hex()
-                )
-            raise NoValidError("Device unkown")
-
-        firmware = sensor_type
-
-        # Mi Scale data length = message length
-        # -all bytes before Mi Scale UUID
-        # -3 bytes UUID + ADtype
-        # -1 RSSI (normal, not extended packet only)
-        xdata_length = msg_length - miscale_index - 3 - (0 if is_ext_packet else 1)
-        if xdata_length != (10 if sensor_type == "Mi Scale V1" else 13):
-            raise NoValidError("Xdata length invalid")
-
-        xdata_point = miscale_index + 3
-        xnext_point = xdata_point + xdata_length
-        xvalue = data[xdata_point:xnext_point]
-
-        packet_id = xvalue.hex()
-        try:
-            prev_packet = self.lpacket_ids[miscale_index]
-        except KeyError:
-            # start with empty first packet
-            prev_packet = None, None, None
-        if prev_packet == packet_id:
-            # only process new messages
-            return None, None, None
-        self.lpacket_ids[miscale_index] = packet_id
-
-        result = {
-            "rssi": rssi,
-            "mac": ''.join('{:02X}'.format(x) for x in miscale_mac[:]),
-            "type": sensor_type,
-            "packet": packet_id,
-            "firmware": firmware,
-            "data": True,
-        }
-
-        binary = False
-        measuring = False
-        xvalue_typecode = data[miscale_index + 1:miscale_index + 3]
-
-        resfunc, tbinary, tmeasuring = self._dataobject_dict.get(xvalue_typecode, (None, None, None))
-        if resfunc:
-            binary = binary or tbinary
-            measuring = measuring or tmeasuring
-            result.update(resfunc(xvalue))
-        else:
-            if self.report_unknown:
-                _LOGGER.info(
-                    "UNKNOWN dataobject from Mi Scale DEVICE: %s, MAC: %s, ADV: %s",
-                    sensor_type,
-                    ''.join('{:02X}'.format(x) for x in miscale_mac[:]),
-                    data.hex()
-                )
-        binary = binary and binary_data
-        return result, binary, measuring
-
-    def parse_kegtron(self, data, kegtron_index, is_ext_packet):
-        # parse BLE message in Kegtron format
-        firmware = "Kegtron"
-
-        # check for no BR/EDR + LE General discoverable mode flags
-        advert_start = 29 if is_ext_packet else 14
-        adv_index = data.find(b"\x1E\xFF\xFF", advert_start, advert_start + 3)
-        if adv_index == -1:
-            raise NoValidError("Invalid index")
-
-        # check for BTLE msg size
-        msg_length = data[2] + 3
-        if msg_length != len(data):
-            raise NoValidError("Invalid msg size")
-
-        # extract device type
-        device_type = data[kegtron_index:kegtron_index + 4]
-
-        # check for MAC presence in message and in service data
-        mac_index = adv_index - 14 if is_ext_packet else adv_index
-        kegtron_mac_reversed = data[mac_index - 7:mac_index - 1]
-
-        # check for MAC presence in whitelist, if needed
-        if self.discovery is False and kegtron_mac_reversed not in self.whitelist:
-            return None, None, None
-        packet_id = "no packed id"
-
-        # extract RSSI byte
-        rssi_index = 18 if is_ext_packet else msg_length - 1
-        (rssi,) = struct.unpack("<b", data[rssi_index:rssi_index + 1])
-
-        # strange positive RSSI workaround
-        if rssi > 0:
-            rssi = -rssi
-        try:
-            sensor_type, binary_data = KEGTRON_TYPE_DICT[device_type]
-        except KeyError:
-            if self.report_unknown:
-                _LOGGER.info(
-                    "BLE ADV from UNKNOWN: RSSI: %s, MAC: %s, ADV: %s",
-                    rssi,
-                    ''.join('{:02X}'.format(x) for x in kegtron_mac_reversed[::-1]),
-                    data.hex()
-                )
-            raise NoValidError("Device unkown")
-
-        # kegtron data length = message length
-        #     -all bytes before Kegtron UUID
-        #     -1 Len byte
-        #     -3 bytes kegtron UUID + ADtype
-        #     -1 RSSI (normal, not extended packet only)
-        xdata_length = msg_length - kegtron_index - (4 if is_ext_packet else 5)
-        if xdata_length != 27:
-            raise NoValidError("Xdata length invalid")
-
-        xdata_point = kegtron_index + 4
-
-        # check if kegtron data start and length is valid
-        if xdata_length != len(data[xdata_point:-1]):
-            raise NoValidError("Invalid data length")
-        result = {
-            "rssi": rssi,
-            "mac": ''.join('{:02X}'.format(x) for x in kegtron_mac_reversed[::-1]),
-            "type": sensor_type,
-            "packet": packet_id,
-            "firmware": firmware,
-            "data": True,
-        }
-        binary = False
-        measuring = False
-
-        xvalue_typecode = data[xdata_point - 2:xdata_point]
-        xvalue_length = xdata_length
-        xnext_point = xdata_point + xvalue_length
-        xvalue = data[xdata_point:xnext_point]
-        resfunc, tbinary, tmeasuring = self._dataobject_dict.get(xvalue_typecode, (None, None, None))
-        if resfunc:
-            binary = binary or tbinary
-            measuring = measuring or tmeasuring
-            result.update(resfunc(xvalue))
-        else:
-            if self.report_unknown:
-                _LOGGER.info(
-                    "UNKNOWN dataobject from DEVICE: %s, MAC: %s, ADV: %s",
-                    sensor_type,
-                    ''.join('{:02X}'.format(x) for x in kegtron_mac_reversed[::-1]),
-                    data.hex()
-                )
-        binary = binary and binary_data
-        return result, binary, measuring
-
-
-class NoValidError(Exception):
-    pass
