+++ custom_components/ble_monitor/const.py
@@ -52,56 +52,6 @@ CONF_TMAX = 60.0
 CONF_HMIN = 0.0
 CONF_HMAX = 99.9
 
-# Xiaomi sensor types dictionary for adv parser
-#                              binary?
-XIAOMI_TYPE_DICT = {
-    b'\xAA\x01': ("LYWSDCGQ", False),
-    b'\x47\x03': ("CGG1", False),
-    b'\x48\x0B': ("CGG1-ENCRYPTED", False),
-    b'\x6F\x06': ("CGDK2", False),
-    b'\x5B\x04': ("LYWSD02", False),
-    b'\x5B\x05': ("LYWSD03MMC", False),
-    b'\x76\x05': ("CGD1", False),
-    b'\xd3\x06': ("MHO-C303", False),
-    b'\x87\x03': ("MHO-C401", False),
-    b'\xDF\x02': ("JQJCY01YM", False),
-    b'\x98\x00': ("HHCCJCY01", False),
-    b'\xBC\x03': ("GCLS002", False),
-    b'\x5D\x01': ("HHCCPOT002", False),
-    b'\x0A\x04': ("WX08ZM", True),
-    b'\x8B\x09': ("MCCGQ02HL", True),
-    b'\xD6\x03': ("CGH1", True),
-    b'\x83\x00': ("YM-K1501", True),
-    b'\x13\x01': ("YM-K1501EU", True),
-    b'\x5C\x04': ("V-SK152", True),
-    b'\x63\x08': ("SJWS01LM", True),
-    b'\xF6\x07': ("MJYD02YL", True),
-    b'\xDD\x03': ("MUE4094RT", True),
-    b'\x8D\x0A': ("RTCGQ02LM", True),
-    b'\x83\x0A': ("CGPR1", True),
-    b'\xDB\x00': ("MMC-T201-1", False),
-    b'\xBF\x07': ("YLAI003", False),
-}
-
-QINGPING_TYPE_DICT = {
-    b'\x08\x07': ("CGG1", False),
-    b'\x08\x09': ("CGP1W", False),
-    b'\x08\x0C': ("CGD1", False),
-}
-
-ATC_TYPE_DICT = {
-    b'\x1A\x18': ("LYWSD03MMC", False),
-}
-
-MISCALE_TYPE_DICT = {
-    b'\x1D\x18': ("Mi Scale V1", True),
-    b'\x1B\x18': ("Mi Scale V2", True),
-}
-
-KEGTRON_TYPE_DICT = {
-    b'\x1E\xFF\xFF\xFF': ("Kegtron", False),
-}
-
 # Sensor type indexes dictionary for sensor platform
 # Sensor:
 # T  = Temperature
@@ -116,8 +66,6 @@ KEGTRON_TYPE_DICT = {
 # W  = Weight
 # NW = Non-stabilized Weight
 # Im = Impedance
-# Ks = Keg size
-# Vs = Volume start
 # Vd = Volume dispensed
 # V  = Voltage
 # B  = Battery
@@ -130,39 +78,40 @@ KEGTRON_TYPE_DICT = {
 # Mn = Motion
 # WR = Weight Removed
 # B  = Battery
-#                                  sensor                                        binary sensor
-# Measurement type      [T  H  M  P  C  I  F  Cn Bu W  NW Im Ks Vs Vd V  B]  [Sw Op L  Mo Mn WR B]  (start from 0, 9 - no data)
+#                                  sensor                                   binary sensor
+# Measurement type      [T  H  M  P  C  I  F  Cn Bu W  NW Im Vd V  B]  [Sw Op L  Mo Mn WR B] (start from 0, 9 - no data)
 MMTS_DICT = {
-    'LYWSDCGQ'       : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
-    'CGG1'           : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
-    'CGG1-ENCRYPTED' : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
-    'CGDK2'          : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
-    'LYWSD02'        : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
-    'LYWSD03MMC'     : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 3, 2], [9, 9, 9, 9, 9, 9, 9]],
-    'CGD1'           : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
-    'CGP1W'          : [[0, 1, 9, 2, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 3], [9, 9, 9, 9, 9, 9, 9]],
-    'MHO-C401'       : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
-    'MHO-C303'       : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
-    'JQJCY01YM'      : [[0, 1, 9, 9, 9, 9, 2, 9, 9, 9, 9, 9, 9, 9, 9, 9, 3], [9, 9, 9, 9, 9, 9, 9]],
-    'HHCCJCY01'      : [[0, 9, 1, 9, 2, 3, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [9, 9, 9, 9, 9, 9, 9]],
-    'GCLS002'        : [[0, 9, 1, 9, 2, 3, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [9, 9, 9, 9, 9, 9, 9]],
-    'HHCCPOT002'     : [[9, 9, 0, 9, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [9, 9, 9, 9, 9, 9, 9]],
-    'WX08ZM'         : [[9, 9, 9, 9, 9, 9, 9, 0, 9, 9, 9, 9, 9, 9, 9, 9, 1], [0, 9, 9, 9, 9, 9, 1]],
-    'MCCGQ02HL'      : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0], [9, 0, 1, 9, 9, 9, 2]],
-    'CGH1'           : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0], [9, 0, 9, 9, 9, 9, 1]],
-    'YM-K1501'       : [[0, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [0, 9, 9, 9, 9, 9, 9]],
-    'YM-K1501EU'     : [[0, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [0, 9, 9, 9, 9, 9, 9]],
-    'V-SK152'        : [[0, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [0, 9, 9, 9, 9, 9, 9]],
-    'SJWS01LM'       : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0], [9, 9, 9, 0, 9, 9, 1]],
-    'MJYD02YL'       : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0], [9, 9, 0, 9, 1, 9, 2]],
-    'MUE4094RT'      : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [9, 9, 9, 9, 0, 9, 9]],
-    'RTCGQ02LM'      : [[9, 9, 9, 9, 9, 9, 9, 9, 0, 9, 9, 9, 9, 9, 9, 9, 1], [9, 9, 0, 9, 1, 9, 2]],
-    'CGPR1'          : [[9, 9, 9, 9, 9, 0, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 1], [9, 9, 9, 9, 0, 9, 1]],
-    'MMC-T201-1'     : [[0, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 1], [9, 9, 9, 9, 9, 9, 9]],
-    'YLAI003'        : [[9, 9, 9, 9, 9, 9, 9, 9, 0, 9, 9, 9, 9, 9, 9, 9, 1], [9, 9, 9, 9, 9, 9, 9]],
-    'Mi Scale V1'    : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 0, 1, 9, 9, 9, 9, 9, 9], [9, 9, 9, 9, 9, 0, 9]],
-    'Mi Scale V2'    : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 0, 1, 2, 9, 9, 9, 9, 9], [9, 9, 9, 9, 9, 0, 9]],
-    'Kegtron'        : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0, 1, 2, 9, 9], [9, 9, 9, 9, 9, 9, 9]],
+    'LYWSDCGQ'       : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
+    'CGG1'           : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
+    'CGG1-ENCRYPTED' : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
+    'CGDK2'          : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
+    'LYWSD02'        : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
+    'LYWSD03MMC'     : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 3, 2], [9, 9, 9, 9, 9, 9, 9]],
+    'CGD1'           : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
+    'CGP1W'          : [[0, 1, 9, 2, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 3], [9, 9, 9, 9, 9, 9, 9]],
+    'MHO-C401'       : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
+    'MHO-C303'       : [[0, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 2], [9, 9, 9, 9, 9, 9, 9]],
+    'JQJCY01YM'      : [[0, 1, 9, 9, 9, 9, 2, 9, 9, 9, 9, 9, 9, 9, 3], [9, 9, 9, 9, 9, 9, 9]],
+    'HHCCJCY01'      : [[0, 9, 1, 9, 2, 3, 9, 9, 9, 9, 9, 9, 9, 9, 9], [9, 9, 9, 9, 9, 9, 9]],
+    'GCLS002'        : [[0, 9, 1, 9, 2, 3, 9, 9, 9, 9, 9, 9, 9, 9, 9], [9, 9, 9, 9, 9, 9, 9]],
+    'HHCCPOT002'     : [[9, 9, 0, 9, 1, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [9, 9, 9, 9, 9, 9, 9]],
+    'WX08ZM'         : [[9, 9, 9, 9, 9, 9, 9, 0, 9, 9, 9, 9, 9, 9, 1], [0, 9, 9, 9, 9, 9, 1]],
+    'MCCGQ02HL'      : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0], [9, 0, 1, 9, 9, 9, 2]],
+    'CGH1'           : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0], [9, 0, 9, 9, 9, 9, 1]],
+    'YM-K1501'       : [[0, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [0, 9, 9, 9, 9, 9, 9]],
+    'YM-K1501EU'     : [[0, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [0, 9, 9, 9, 9, 9, 9]],
+    'V-SK152'        : [[0, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [0, 9, 9, 9, 9, 9, 9]],
+    'SJWS01LM'       : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0], [9, 9, 9, 0, 9, 9, 1]],
+    'MJYD02YL'       : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0], [9, 9, 0, 9, 1, 9, 2]],
+    'MUE4094RT'      : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9], [9, 9, 9, 9, 0, 9, 9]],
+    'RTCGQ02LM'      : [[9, 9, 9, 9, 9, 9, 9, 9, 0, 9, 9, 9, 9, 9, 1], [9, 9, 0, 9, 1, 9, 2]],
+    'CGPR1'          : [[9, 9, 9, 9, 9, 0, 9, 9, 9, 9, 9, 9, 9, 9, 1], [9, 9, 9, 9, 0, 9, 1]],
+    'MMC-T201-1'     : [[0, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 1], [9, 9, 9, 9, 9, 9, 9]],
+    'YLAI003'        : [[9, 9, 9, 9, 9, 9, 9, 9, 0, 9, 9, 9, 9, 9, 1], [9, 9, 9, 9, 9, 9, 9]],
+    'Mi Scale V1'    : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 0, 1, 9, 9, 9, 9], [9, 9, 9, 9, 9, 0, 9]],
+    'Mi Scale V2'    : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 0, 1, 2, 9, 9, 9], [9, 9, 9, 9, 9, 0, 9]],
+    'Kegtron KT-100' : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0, 9, 9], [9, 9, 9, 9, 9, 9, 9]],
+    'Kegtron KT-200' : [[9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 0, 9, 9], [9, 9, 9, 9, 9, 9, 9]],
 }
 
 KETTLES = ('YM-K1501', 'YM-K1501EU', 'V-SK152')
@@ -198,5 +147,6 @@ MANUFACTURER_DICT = {
     'YLAI003'        : 'Yeelight',
     'Mi Scale V1'    : 'Xiaomi',
     'Mi Scale V2'    : 'Xiaomi',
-    'Kegtron'        : 'Kegtron',
+    'Kegtron KT-100' : 'Kegtron',
+    'Kegtron KT-200' : 'Kegtron',
 }
