+++ custom_components/ble_monitor/manifest.json
@@ -10,5 +10,5 @@
     "@Ernst79",
     "@Magalex2x14"
   ],
-  "version": "1.8.0-beta"
+  "version": "1.8.1-beta"
 }
